internal/core/organization_onboarding/api/get_onboarding_status.go
package api

/**
 * Panther is a scalable, powerful, cloud-native SIEM written in Golang/React.
 * Copyright (C) 2020 Panther Labs Inc
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as
 * published by the Free Software Foundation, either version 3 of the
 * License, or (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */

import (
	"github.com/panther-labs/panther/api/lambda/onboarding/models"
)

// GetOnboardingStatus calls stepFunctionGateway to get onboarding status.
func (API) GetOnboardingStatus(input *models.GetOnboardingStatusInput) (*models.GetOnboardingStatusOutput, error) {
	o, err := stepFunctionGateway.DescribeExecution(input.ExecutionArn)
	if err != nil {
		return nil, err
	}
	return o, nil
}
